move restore caching helpers to new file restore_caching.py
